Removed the unwanted files, also updated the load_data() function
